added basic masteries
I implemented a very barebone version of leveling up masteries. They are not saved between session, but always reset to 0 on logging in. I intend to load the mastery levels from the database on log in. Saving the mastery level should happen right before sending the update to the client.