Mastery level-up and beginnings of learning skills implemented
I added a MASTERIES table to the database and am reading from it when loading the player. I am also inserting into this table when creating a character, so it starts with all masteries at level 0. I am not updating the database when leveling up masteries yet, because very little is getting written to the database until now, and this is a seperate issue.